all-devices-app: Rename NIM to NetworkInfrastructureManager and refactor to Device-is-Delegate pattern
