+++ examples/all-devices-app/all-devices-common/devices/enabled_devices_config.h.in
@@ -12,7 +12,7 @@
 #cmakedefine01 ALL_DEVICES_ENABLE_OCCUPANCY_SENSOR
 #cmakedefine01 ALL_DEVICES_ENABLE_CHIME
 #cmakedefine01 ALL_DEVICES_ENABLE_DIMMABLE_LIGHT
-#cmakedefine01 ALL_DEVICES_ENABLE_NIM
+#cmakedefine01 ALL_DEVICES_ENABLE_NETWORK_INFRASTRUCTURE_MANAGER
 #cmakedefine01 ALL_DEVICES_ENABLE_ON_OFF_LIGHT
 #cmakedefine01 ALL_DEVICES_ENABLE_SPEAKER
 #cmakedefine01 ALL_DEVICES_ENABLE_SOIL_SENSOR
