+++ examples/all-devices-app/all-devices-common/devices/network-infrastructure-manager/BUILD.gn
@@ -16,11 +16,10 @@
 import("//build_overrides/build.gni")
 import("//build_overrides/chip.gni")
 
-source_set("nim") {
+source_set("network-infrastructure-manager") {
   sources = [
-    "LoggingBorderRouterDelegate.h",
-    "NimDevice.cpp",
-    "NimDevice.h",
+    "NetworkInfrastructureManagerDevice.cpp",
+    "NetworkInfrastructureManagerDevice.h",
   ]
 
   public_deps = [
