+++ src/app/tests/suites/certification/Test_TC_TBRM_2_4.yaml
@@ -128,6 +128,6 @@ tests:
 CI:
     - name: network-manager
       app: network-manager
-    - name: nim-device
+    - name: network-infrastructure-manager
       app: all-devices
-      args: ["--device", "nim"]
+      args: ["--device", "network-infrastructure-manager"]
