+++ src/app/tests/suites/certification/Test_TC_THNETDIR_2_3.yaml
@@ -95,6 +95,6 @@ tests:
 CI:
     - name: network-manager
       app: network-manager
-    - name: nim-device
+    - name: network-infrastructure-manager
       app: all-devices
-      args: ["--device", "nim"]
+      args: ["--device", "network-infrastructure-manager"]
