+++ src/app/tests/suites/certification/Test_TC_WIFINM_2_1.yaml
@@ -67,6 +67,6 @@ tests:
 CI:
     - name: network-manager
       app: network-manager
-    - name: nim-device
+    - name: network-infrastructure-manager
       app: all-devices
-      args: ["--device", "nim"]
+      args: ["--device", "network-infrastructure-manager"]
